Copied Index File from ITFN Repository
I originally created the ITFnd100-Mod06 repository during the lecture but later learned that, for the homework, the repository should be named IntroToProg-Python-Mod06.  I uploaded the word document and the code to the new repository but forgot to upload the index file.  I am fixing this by adding the index file to this repository and I will be deleting the old repository.